requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
langchain==0.3.13
langchain-openai==0.2.14
langchain-core==0.3.28
langgraph==0.2.60
python-dotenv==1.0.1
